Layout: add navbar and move component to own folder
